remove old project model
